requirements.txt
bleak>=0.20.0
Pillow>=9.0.0
numpy>=1.20.0